requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.0.1
requests==2.31.0
openai==1.3.0  # For OpenRouter API
ollama==0.1.7  # For Ollama fallback
reportlab==4.0.4  # For PDF generation
weasyprint==61.0  # For HTML to PDF
beautifulsoup4==4.12.2
lxml==4.9.3
